+++ Changelog.md
@@ -1,5 +1,6 @@
 ## 13.05.2025
 1. Add support for fetching blocks by number and dynamic time
+2. Add support for skipped blocks
 
 ## 18.04.2025
 1. Feeder: added int blockDto: "From", "TransactionHash" to each receipt